refactor: Seperate sentry init functions to improve index cleanliness
