+++ eslint.config.mjs
@@ -64,6 +64,6 @@ export default defineConfig([globalIgnores([
         "markers": ["!"]
       }
     ],
-    "object-curly-spacing": ["error", "always"]
+    "object-curly-spacing": ["error", "always"],
   },
 }]);
